+++ Src/FluentAssertions/Primitives/DateTimeOffsetRangeAssertions.cs
@@ -23,7 +23,6 @@ public class DateTimeOffsetRangeAssertions<TAssertions>
     #region Private Definitions
 
     private readonly TAssertions parentAssertions;
-    private readonly AssertionChain assertionChain;
     private readonly TimeSpanPredicate predicate;
 
     private readonly Dictionary<TimeSpanCondition, TimeSpanPredicate> predicates = new()
@@ -46,12 +45,17 @@ public class DateTimeOffsetRangeAssertions<TAssertions>
         TimeSpan timeSpan)
     {
         this.parentAssertions = parentAssertions;
-        this.assertionChain = assertionChain;
+        CurrentAssertionChain = assertionChain;
         this.subject = subject;
         this.timeSpan = timeSpan;
 
         predicate = predicates[condition];
     }
+
+    /// <summary>
+    /// Provides access to the <see cref="AssertionChain"/> that this assertion class was initialized with.
+    /// </summary>
+    public AssertionChain CurrentAssertionChain { get; }
 
     /// <summary>
     /// Asserts that a <see cref="DateTimeOffset"/> occurs a specified amount of time before another <see cref="DateTimeOffset"/>.
@@ -69,17 +73,17 @@ public class DateTimeOffsetRangeAssertions<TAssertions>
     public AndConstraint<TAssertions> Before(DateTimeOffset target,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        assertionChain
+        CurrentAssertionChain
             .ForCondition(subject.HasValue)
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:the date and time} to be " + predicate.DisplayText +
                 " {0} before {1}{reason}, but found a <null> DateTime.", timeSpan, target);
 
-        if (assertionChain.Succeeded)
+        if (CurrentAssertionChain.Succeeded)
         {
             TimeSpan actual = target - subject.Value;
 
-            assertionChain
+            CurrentAssertionChain
                 .ForCondition(predicate.IsMatchedBy(actual, timeSpan))
                 .BecauseOf(because, becauseArgs)
                 .FailWith(
@@ -107,17 +111,17 @@ public class DateTimeOffsetRangeAssertions<TAssertions>
     public AndConstraint<TAssertions> After(DateTimeOffset target,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        assertionChain
+        CurrentAssertionChain
             .ForCondition(subject.HasValue)
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:the date and time} to be " + predicate.DisplayText +
                 " {0} after {1}{reason}, but found a <null> DateTime.", timeSpan, target);
 
-        if (assertionChain.Succeeded)
+        if (CurrentAssertionChain.Succeeded)
         {
             TimeSpan actual = subject.Value - target;
 
-            assertionChain
+            CurrentAssertionChain
                 .ForCondition(predicate.IsMatchedBy(actual, timeSpan))
                 .BecauseOf(because, becauseArgs)
                 .FailWith(