+++ Src/FluentAssertions/Specialized/TaskCompletionSourceAssertions.cs
@@ -12,21 +12,25 @@ namespace FluentAssertions.Specialized;
 #if NET6_0_OR_GREATER
 public class TaskCompletionSourceAssertions : TaskCompletionSourceAssertionsBase
 {
-    private readonly AssertionChain assertionChain;
     private readonly TaskCompletionSource subject;
 
     public TaskCompletionSourceAssertions(TaskCompletionSource tcs, AssertionChain assertionChain)
         : this(tcs, assertionChain, new Clock())
     {
-        this.assertionChain = assertionChain;
+        CurrentAssertionChain = assertionChain;
     }
 
     public TaskCompletionSourceAssertions(TaskCompletionSource tcs, AssertionChain assertionChain, IClock clock)
         : base(clock)
     {
         subject = tcs;
-        this.assertionChain = assertionChain;
+        CurrentAssertionChain = assertionChain;
     }
+
+    /// <summary>
+    /// Provides access to the <see cref="AssertionChain"/> that this assertion class was initialized with.
+    /// </summary>
+    public AssertionChain CurrentAssertionChain { get; }
 
     /// <summary>
     /// Asserts that the <see cref="Task"/> of the current <see cref="TaskCompletionSource"/> will complete within the specified time.
@@ -42,15 +46,15 @@ public class TaskCompletionSourceAssertions : TaskCompletionSourceAssertionsBase
     public async Task<AndConstraint<TaskCompletionSourceAssertions>> CompleteWithinAsync(
         TimeSpan timeSpan, [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        assertionChain
+        CurrentAssertionChain
             .ForCondition(subject is not null)
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context} to complete within {0}{reason}, but found <null>.", timeSpan);
 
-        if (assertionChain.Succeeded)
+        if (CurrentAssertionChain.Succeeded)
         {
             bool completesWithinTimeout = await CompletesWithinTimeoutAsync(subject!.Task, timeSpan);
-            assertionChain
+            CurrentAssertionChain
                 .ForCondition(completesWithinTimeout)
                 .BecauseOf(because, becauseArgs)
                 .FailWith("Expected {context:task} to complete within {0}{reason}.", timeSpan);
@@ -73,15 +77,15 @@ public class TaskCompletionSourceAssertions : TaskCompletionSourceAssertionsBase
     public async Task<AndConstraint<TaskCompletionSourceAssertions>> NotCompleteWithinAsync(
         TimeSpan timeSpan, [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        assertionChain
+        CurrentAssertionChain
             .ForCondition(subject is not null)
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context} to not complete within {0}{reason}, but found <null>.", timeSpan);
 
-        if (assertionChain.Succeeded)
+        if (CurrentAssertionChain.Succeeded)
         {
             bool completesWithinTimeout = await CompletesWithinTimeoutAsync(subject!.Task, timeSpan);
-            assertionChain
+            CurrentAssertionChain
                 .ForCondition(!completesWithinTimeout)
                 .BecauseOf(because, becauseArgs)
                 .FailWith("Expected {context:task} to not complete within {0}{reason}.", timeSpan);
@@ -94,21 +98,25 @@ public class TaskCompletionSourceAssertions : TaskCompletionSourceAssertionsBase
 
 public class TaskCompletionSourceAssertions<T> : TaskCompletionSourceAssertionsBase
 {
-    private readonly AssertionChain assertionChain;
     private readonly TaskCompletionSource<T> subject;
 
     public TaskCompletionSourceAssertions(TaskCompletionSource<T> tcs, AssertionChain assertionChain)
         : this(tcs, assertionChain, new Clock())
     {
-        this.assertionChain = assertionChain;
+        CurrentAssertionChain = assertionChain;
     }
 
     public TaskCompletionSourceAssertions(TaskCompletionSource<T> tcs, AssertionChain assertionChain, IClock clock)
         : base(clock)
     {
         subject = tcs;
-        this.assertionChain = assertionChain;
+        CurrentAssertionChain = assertionChain;
     }
+
+    /// <summary>
+    /// Provides access to the <see cref="AssertionChain"/> that this assertion class was initialized with.
+    /// </summary>
+    public AssertionChain CurrentAssertionChain { get; }
 
     /// <summary>
     /// Asserts that the <see cref="Task"/> of the current <see cref="TaskCompletionSource{T}"/> will complete within the specified time.
@@ -124,16 +132,16 @@ public class TaskCompletionSourceAssertions<T> : TaskCompletionSourceAssertionsB
     public async Task<AndWhichConstraint<TaskCompletionSourceAssertions<T>, T>> CompleteWithinAsync(
         TimeSpan timeSpan, [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        assertionChain
+        CurrentAssertionChain
             .ForCondition(subject is not null)
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context} to complete within {0}{reason}, but found <null>.", timeSpan);
 
-        if (assertionChain.Succeeded)
+        if (CurrentAssertionChain.Succeeded)
         {
             bool completesWithinTimeout = await CompletesWithinTimeoutAsync(subject!.Task, timeSpan);
 
-            assertionChain
+            CurrentAssertionChain
                 .ForCondition(completesWithinTimeout)
                 .BecauseOf(because, becauseArgs)
                 .FailWith("Expected {context:task} to complete within {0}{reason}.", timeSpan);
@@ -141,7 +149,7 @@ public class TaskCompletionSourceAssertions<T> : TaskCompletionSourceAssertionsB
 #pragma warning disable CA1849 // Call async methods when in an async method
             T result = subject.Task.IsCompleted ? subject.Task.Result : default;
 #pragma warning restore CA1849 // Call async methods when in an async method
-            return new AndWhichConstraint<TaskCompletionSourceAssertions<T>, T>(this, result, assertionChain, ".Result");
+            return new AndWhichConstraint<TaskCompletionSourceAssertions<T>, T>(this, result, CurrentAssertionChain, ".Result");
         }
 
         return new AndWhichConstraint<TaskCompletionSourceAssertions<T>, T>(this, default(T));
@@ -161,16 +169,16 @@ public class TaskCompletionSourceAssertions<T> : TaskCompletionSourceAssertionsB
     public async Task<AndConstraint<TaskCompletionSourceAssertions<T>>> NotCompleteWithinAsync(
         TimeSpan timeSpan, [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        assertionChain
+        CurrentAssertionChain
             .ForCondition(subject is not null)
             .BecauseOf(because, becauseArgs)
             .FailWith("Did not expect {context} to complete within {0}{reason}, but found <null>.", timeSpan);
 
-        if (assertionChain.Succeeded)
+        if (CurrentAssertionChain.Succeeded)
         {
             bool completesWithinTimeout = await CompletesWithinTimeoutAsync(subject!.Task, timeSpan);
 
-            assertionChain
+            CurrentAssertionChain
                 .ForCondition(!completesWithinTimeout)
                 .BecauseOf(because, becauseArgs)
                 .FailWith("Did not expect {context:task} to complete within {0}{reason}.", timeSpan);